# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a crypto wallet api",
        "title": "Crypto Wallet API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "info@info.info"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/hello": {
            "get": {
                "description": "Method to say hello",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hello"
                ],
                "summary": "Method to say hello",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/wallets": {
            "post": {
                "description": "Method for create wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Create wallet info",
                "parameters": [
                    {
                        "description": "The input for create wallet",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallet.CreateWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.CreateWalletResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Method for Delete wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Delete wallet info",
                "parameters": [
                    {
                        "description": "The input for delete wallet",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallet.DeleteWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/wallets/import": {
            "post": {
                "description": "Method for import wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Import wallet info",
                "parameters": [
                    {
                        "description": "The input for import wallet",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallet.ImportWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.ImportWalletResponse"
                        }
                    }
                }
            }
        },
        "/wallets/{id}": {
            "get": {
                "description": "Method for get wallet info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get wallet info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pass phrase",
                        "name": "passphrase",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallet.GetWalletInfoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "wallet.CreateWalletRequest": {
            "type": "object",
            "properties": {
                "passphrase": {
                    "type": "string"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
        "wallet.CreateWalletResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "object",
                    "properties": {
                        "algorithm": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "version": {
                                    "type": "integer"
                                }
                            }
                        },
                        "meta": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wallet.Meta"
                            }
                        },
                        "publicKey": {
                            "type": "string"
                        }
                    }
                },
                "wallet": {
                    "type": "object",
                    "properties": {
                        "filePath": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "recoveryPhrase": {
                            "type": "string"
                        },
                        "version": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "wallet.DeleteWalletRequest": {
            "type": "object",
            "properties": {
                "wallet": {
                    "type": "string"
                }
            }
        },
        "wallet.GetWalletInfoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "wallet.ImportWalletRequest": {
            "type": "object",
            "properties": {
                "passphrase": {
                    "type": "string"
                },
                "recoveryPhrase": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
        "wallet.ImportWalletResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "object",
                    "properties": {
                        "algorithm": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "version": {
                                    "type": "integer"
                                }
                            }
                        },
                        "meta": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/wallet.Meta"
                            }
                        },
                        "publicKey": {
                            "type": "string"
                        }
                    }
                },
                "wallet": {
                    "type": "object",
                    "properties": {
                        "filePath": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "version": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "wallet.Meta": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}